#60 Rename module (generic → mixins)
