+++ pandas_dataclasses/__init__.py
@@ -1,6 +1,6 @@
 # type: ignore
 from . import core
-from .core.generic import *
+from .core.mixins import *
 from .core.parsers import *
 from .core.typing import *
