+++ tests/test_frame.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 # dependencies
 import numpy as np
 import pandas as pd
-from pandas_dataclasses.core.generic import As
+from pandas_dataclasses.core.mixins import As
 from pandas_dataclasses.core.typing import Attr, Data, Index
 from typing_extensions import Annotated as Ann
